added support for using angular's attributes object as override object with data type conversion

+++ defaults-factory.js
@@ -29,8 +29,8 @@ angular.module('nag.core')
           _.forEach(finalOptions, function(value, key) {
             if(options[key] === 'true' || options[key] === 'false') {
               temp[key] = options[key] === 'true' ? true : false;
-            } else if(!isNaN(parseInt(options[key]))) {
-              temp[key] = parseInt(options[key]);
+            } else if(!isNaN(parseFloat(options[key]))) {
+              temp[key] = parseFloat(options[key]);
             } else if(options[key] !== undefined) {
               temp[key] = options[key];
             }
